chore: refactored next plugin styles
