+++ front/src/pages/LessonEdit/LessonEdit.jsx
@@ -11,8 +11,8 @@ import Header from '@sb-ui/components/molecules/Header';
 import { createLesson, getLesson, putLesson } from '@sb-ui/utils/api/v1/lesson';
 import { Statuses } from '@sb-ui/pages/TeacherHome/LessonsDashboard/constants';
 import { LESSONS_EDIT } from '@sb-ui/utils/paths';
-import Next from '@sb-ui/utils/next-plugin/next';
-import Undo from '@sb-ui/utils/undo-plugin';
+import Next from '@sb-ui/utils/editorjs/next-plugin';
+import Undo from '@sb-ui/utils/editorjs/undo-plugin';
 import * as S from './LessonEdit.styled';
 
 const { TextArea } = Input;
@@ -79,7 +79,7 @@ const LessonEdit = () => {
       },
       plugins: [],
     });
-  }, []);
+  }, [t]);
 
   useEffect(() => {
     if (inputTitle.current && !isLoading && !lessonData?.lesson.name) {
