+++ front/src/utils/editorjs/next-plugin/index.js
@@ -1,0 +1,1 @@
+export { default } from './next';
