+++ front/src/utils/editorjs/next-plugin/next.css
@@ -1,0 +1,18 @@
+.next-tool__wrapper {
+  margin: 1rem 0;
+}
+
+.next-tool__base {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 40px;
+  background: #ffffff;
+  border: 1px dashed #d9d9d9;
+  box-sizing: border-box;
+  box-shadow: 0 2px 0 rgba(0, 0, 0, 0.016);
+  border-radius: 2px;
+  background: inherit;
+}
