+++ front/src/utils/editorjs/undo-plugin/index.js
