--- front/src/utils/next-plugin/next.css
@@ -1,13 +1,0 @@
-.next-tool__base{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 40px;
-    background: #FFFFFF;
-    border: 1px dashed #D9D9D9;
-    box-sizing: border-box;
-    box-shadow: 0 2px 0 rgba(0, 0, 0, 0.016);
-    border-radius: 2px;
-}
